requirements.txt
nemo_toolkit[asr]>=2.0.0
soundfile>=0.12.1
numba<0.63
